Add upcoming service list
