--- src/components/molecules/ServiceGeneralCard/ServiceCardData.js
@@ -1,26 +1,0 @@
-const serviceCardData = () => {
-  return [
-    {
-      title: 'Fecha: 04/02/2022' /* + ${prop.fecha} */,
-      size: 'h7',
-      color: 'text',
-    },
-    {
-      title: 'Horario: 4:00pm',
-      size: 'h7',
-      color: 'text',
-    },
-    {
-      title: 'Proveedor de Servicio: Angel Abraham Flores',
-      size: 'h7',
-      color: 'text',
-    },
-    {
-      title: 'Cliente: Marecela Rodriguez',
-      size: 'h7',
-      color: 'text',
-    },
-  ];
-};
-
-export default serviceCardData;
