+++ src/components/organisms/index.js
@@ -1,3 +1,4 @@
 export {CardOrganism} from './CardOrganism';
 export {ProviderCardDets} from './ProviderCardDets';
 export {ServiceClosure} from './ServiceClousure';
+export {Services} from './Services';
